% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_info.R
\name{compound_info}
\alias{compound_info}
\title{Compound info}
\usage{
compound_info(kegg_content)
}
\arguments{
\item{kegg_content}{The content of a KEGG query}
}
\value{
A tibble.
}
\description{
Given a KEGG query obtained using the \code{\link(kegg_query)} function, returns information on a chemical compound. Returned elements are drug ids for compound (when available), pathways the compound is involved in, chemical formula, and external ids (CAS, PubChem, ChEBI, and DrugBank)  for compound.
}
