% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG}
\alias{DRUG}
\title{DRUG}
\format{A list of tibbles:
\describe{
  \item{drug_group}{A tibble mapping KEGG drug IDs to KEGG drug group IDs, group name, and drug name}
  \item{drug_uses}{A tibble mapping KEGG drug IDs to known uses for the drugs (efficacy) and drug name}
  \item{disease_indication}{A tibble mapping KEGG drug IDs to KEGG disease IDs, indication name, and drug name}
  \item{targets}{A tibble mapping KEGG drug IDs to gene Entrez IDs, gene symbols, and drug name}
  \item{pathways}{A tibble mapping KEGG drug IDs to KEGG pathway IDs, pathway name, and drug name}
  \item{preparation}{A tibble mapping KEGG drug IDs to drug name and information if record is a known drug mixture (mixture == 1) or not (mixture == 0).}
}}
\source{
\url{https://www.kegg.jp}
}
\usage{
DRUG
}
\description{
Data from KEGG, collected for drugs. Data was parsed with the retrieve_data function.
}
\keyword{datasets}
