% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUGBANK}
\alias{DRUGBANK}
\title{DrugBank}
\format{A list of tibbles:
\describe{
  \item{general}{}
  \item{category}{}
  \item{pathway}{}
  \item{status}{}
  \item{external_ids}{}
}}
\source{
\url{https://www.drugbank.ca}
}
\usage{
DRUGBANK
}
\description{
Data from DrugBank.
}
\keyword{datasets}
