% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugminer.R
\docType{package}
\name{drugminer}
\alias{drugminer}
\alias{drugminer-package}
\title{DrugMineR}
\description{
This package performs a search and annotation of a given compound of interest based on queries done against the KEGG or PubChem REST APIs.
}
