% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{Extract info}
\usage{
extract_info(kegg_content)
}
\arguments{
\item{kegg_content}{REST API content for a KEGG query on a drug or compound.}
}
\value{
Information on compound or drug, including chemical formula, external ids, disease indications, or pathways involved. Please refer to the \link{\code{drug_info}} or \link{\code(compound_info)} for details.
}
\description{
Given the output of a POST request to KEGG REST API, returns information on either drug or compound queried using the \link{\code{drug_info}} or \link{\code{compound_info}}, respectively. The input to the function comes from the \link{\code{kegg_query}} output.
}
