% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG}
\alias{KEGG}
\title{KEGG}
\format{A list of tibbles:
\describe{
  \item{drug}{A tibble with KEGG drug IDs and drug names}
  \item{drug_group}{A tibble with KEGG drug group IDs and drug group names}
  \item{disease}{A tibble with KEGG disease IDs and disease names}
  \item{pathway}{A tibble with KEGG pathway IDs and pathway names}
  \item{synonyms}{A tibble with KEGG drug IDs, drug names, and drug synonyms}
  \item{compound}{A tibble withe KEGG compound IDs and compound name}
}}
\source{
\url{https://www.kegg.jp}
}
\usage{
KEGG
}
\description{
Data from KEGG. Data was parsed with the retrieve_data function.
}
\keyword{datasets}
