% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_processing.R
\name{kegg_processing}
\alias{kegg_processing}
\title{KEGG processing}
\usage{
kegg_processing(save_dir)
}
\arguments{
\item{save_dir}{User defined save directory for parsed data}
}
\value{
KEGG and FRUG data frames.
}
\description{
Wrapper function for the \code{\link{retrieve_data}} function, building the `KEGG` and `DRUG` data frames if needed (`DrugMineR` comes pre-packaged with these data frames.)
}
