% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name2cid.R
\name{name2cid}
\alias{name2cid}
\title{Find PubChem CID from compound name}
\usage{
name2cid(query)
}
\arguments{
\item{query}{A compound name (as a string)}
}
\value{
A tibble with name and corresponding PubChem CID
}
\description{
Function converts a name query into a PubChem CID using the NCBI's REST API for PubChem.
}
