% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_extractor.R
\name{property_extractor}
\alias{property_extractor}
\title{Property extractor}
\usage{
property_extractor(pubchem_cid)
}
\arguments{
\item{pubchem_cid}{A PubChem CID}
}
\value{
A tibble with CID, molecular formula, molecular weight, canonical SMILES, and InChIKey.
}
\description{
Extracts molecular properties for a compound of interest based on its PubChem CID using the PubChem REST API.
}
