% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem2cas.R
\name{pubchem2cas}
\alias{pubchem2cas}
\title{PubChem to CAS number}
\usage{
pubchem2cas(query)
}
\arguments{
\item{query}{A PubChem CID}
}
\value{
A tibble with PubChem CID and corresponding CAS number when found
}
\description{
Extracts a CAS number for a given PubChem CID.
}
\examples{
pubchem2cas("16038120")
}
