% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve data}
\usage{
retrieve_data(...)
}
\value{
A list of tibbles: compounds, drugs, drug groups, pathways, and disease indications.
}
\description{
Retrieves all data from KEGG using the REST API. This is a no-parameter function.
}
\examples{
kegg_data <- retrieve_data()
}
