% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid2cid.R
\name{sid2cid}
\alias{sid2cid}
\title{Convert PubChem SID to PubChem CID}
\usage{
sid2cid(query)
}
\arguments{
\item{query}{A PubChem SID}
}
\value{
A tibble with PubChem SID and corresponding PubChem CID
}
\description{
Function converts a PubChem SID into a PubChem CID using the NCBI's REST API for PubChem.
}
